Refactoring of `line_profiler.line_profiler`
line_profiler/line_profiler.py[i]::LineProfiler
    Moved common methods like `.__call__()`, `.run[ctx,call]()`, and
    their dependencies to `line_profiler/profiler_mixin.py[i]`

line_profiler/profiler_mixin.py[i]::ByCountProfilerMixin
    New mixin class for common methods shared by
    `line_profiler.line_profiler.LineProfiler` and
    `kernprof.ContextualProfile`

line_profiler/profiler_mixin.py[i]::ByCountProfilerMixin.__call__()
    Replaced self-rolled generator check with
    `inspect.isgeneratorfunction()`
